Pom adjusted to fit JavaFX, basic GUI sort of implemented, but implemented in desired way.
