+++ pom.xml
@@ -93,6 +93,14 @@
                     </execution>
                 </executions>
             </plugin>
+            <plugin>
+                <groupId>com.zenjava</groupId>
+                <artifactId>javafx-maven-plugin</artifactId>
+                <version>8.1.4</version>
+                <configuration>
+                    <mainClass/>
+                </configuration>
+            </plugin>
         </plugins>
     </build>
 
@@ -186,26 +194,10 @@
             <version>3.0.1</version>
         </dependency>
         <dependency>
-            <groupId>org.lwjgl.lwjgl</groupId>
-            <artifactId>lwjgl</artifactId>
-            <version>2.9.3</version>
-        </dependency>
-        <dependency>
             <groupId>com.googlecode.mavennatives</groupId>
             <artifactId>maven-nativedependencies-plugin</artifactId>
             <version>0.0.5</version>
         </dependency>
-        <dependency>
-            <groupId>org.lwjgl.lwjgl</groupId>
-            <artifactId>lwjgl_util</artifactId>
-            <version>2.8.4</version>
-        </dependency>
-        <dependency>
-            <groupId>org.lwjgl.lwjgl</groupId>
-            <artifactId>lwjgl_util_applet</artifactId>
-            <version>2.8.4</version>
-        </dependency>
     </dependencies>
-
 
 </project>
